wazevo(backend): do not have *abi in instruction struct

+++ internal/engine/wazevo/backend/isa/arm64/abi.go
@@ -265,7 +265,7 @@ func (m *machine) lowerCall(si *ssa.Instruction) {
 	}
 	calleeABI := m.compiler.GetFunctionABI(m.compiler.SSABuilder().ResolveSignature(sigID))
 
-	stackSlotSize := calleeABI.AlignedArgResultStackSlotSize()
+	stackSlotSize := int64(calleeABI.AlignedArgResultStackSlotSize())
 	if m.maxRequiredStackSizeForCalls < stackSlotSize+16 {
 		m.maxRequiredStackSizeForCalls = stackSlotSize + 16 // return address frame.
 	}
